On avait oublié ce calcul de nodeCenter dans useOsmRequest
On peut donc maintenant bookmarker des ways
